Reorganize deployment settings commands
