+++ pkg/cmd/pulumi/deployment_settings_config.go
@@ -30,7 +30,6 @@ import (
 	"github.com/pulumi/pulumi/sdk/v3/go/common/diag/colors"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/util/cmdutil"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/util/gitutil"
-	"github.com/pulumi/pulumi/sdk/v3/go/common/util/result"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/workspace"
 	"github.com/spf13/cobra"
 )
@@ -91,80 +90,6 @@ func newDeploymentCmd() *cobra.Command {
 
 	cmd.AddCommand(newDeploymentSettingsCmd())
 	cmd.AddCommand(newDeploymentRunCmd())
-
-	return cmd
-}
-
-func newDeploymentRunCmd() *cobra.Command {
-	// Flags for remote operations.
-	remoteArgs := RemoteArgs{}
-
-	var stack string
-	var suppressPermalink bool
-
-	cmd := &cobra.Command{
-		Use:   "run <operation> [url]",
-		Short: "Launch a deployment job on Pulumi Cloud",
-		Long:  "",
-		Args:  cmdutil.RangeArgs(1, 2),
-		Run: cmdutil.RunResultFunc(func(cmd *cobra.Command, args []string) result.Result {
-			ctx := cmd.Context()
-
-			operation, err := apitype.ParsePulumiOperation(args[0])
-			if err != nil {
-				return result.FromError(err)
-			}
-
-			var url string
-			if len(args) > 1 {
-				url = args[1]
-			}
-
-			display := display.Options{
-				Color: cmdutil.GetGlobalColorization(),
-				// we only suppress permalinks if the user passes true. the default is an empty string
-				// which we pass as 'false'
-				SuppressPermalink: suppressPermalink,
-			}
-
-			project, _, err := readProject()
-			if err != nil && !errors.Is(err, workspace.ErrProjectNotFound) {
-				return result.FromError(err)
-			}
-
-			currentBe, err := currentBackend(ctx, project, display)
-			if err != nil {
-				return result.FromError(err)
-			}
-
-			if !currentBe.SupportsDeployments() {
-				return result.FromError(fmt.Errorf("backends of this type %q do not support deployments",
-					currentBe.Name()))
-			}
-
-			s, err := requireStack(ctx, stack, stackOfferNew|stackSetCurrent, display)
-			if err != nil {
-				return result.FromError(err)
-			}
-
-			if errResult := validateDeploymentFlags(url, remoteArgs); errResult != nil {
-				return errResult
-			}
-
-			return runDeployment(ctx, cmd, display, operation, s.Ref().FullyQualifiedName().String(), url, remoteArgs)
-		}),
-	}
-
-	// Remote flags
-	remoteArgs.applyFlagsForDeploymentCommand(cmd)
-
-	cmd.PersistentFlags().BoolVar(
-		&suppressPermalink, "suppress-permalink", false,
-		"Suppress display of the state permalink")
-
-	cmd.PersistentFlags().StringVarP(
-		&stack, "stack", "s", "",
-		"The name of the stack to operate on. Defaults to the current stack")
 
 	return cmd
 }
@@ -325,154 +250,6 @@ func newDeploymentSettingsInitCmd() *cobra.Command {
 	return cmd
 }
 
-func newDeploymentSettingsPullCmd() *cobra.Command {
-	var stack string
-
-	cmd := &cobra.Command{
-		Use:   "pull",
-		Args:  cmdutil.ExactArgs(0),
-		Short: "Pull the stack's deployment settings from Pulumi Cloud into the deployment.yaml file",
-		Long:  "",
-		Run: cmdutil.RunFunc(func(cmd *cobra.Command, args []string) error {
-			ctx := cmd.Context()
-			d, err := initializeDeploymentSettingsCmd(cmd, stack)
-			if err != nil {
-				return err
-			}
-
-			ds, err := d.Backend.GetStackDeploymentSettings(ctx, d.Stack)
-			if err != nil {
-				return err
-			}
-
-			newStackDeployment := &workspace.ProjectStackDeployment{
-				DeploymentSettings: *ds,
-			}
-
-			err = saveProjectStackDeployment(newStackDeployment, d.Stack)
-			if err != nil {
-				return err
-			}
-
-			fmt.Println("Done")
-
-			return nil
-		}),
-	}
-
-	cmd.PersistentFlags().StringVarP(
-		&stack, "stack", "s", "",
-		"The name of the stack to operate on. Defaults to the current stack")
-
-	return cmd
-}
-
-func newDeploymentSettingsUpdateCmd() *cobra.Command {
-	var stack string
-	var yes bool
-
-	cmd := &cobra.Command{
-		Use:        "up",
-		Aliases:    []string{"update"},
-		SuggestFor: []string{"apply", "deploy", "push"},
-		Args:       cmdutil.ExactArgs(0),
-		Short:      "Update stack deployment settings from deployment.yaml",
-		Long:       "",
-		Run: cmdutil.RunFunc(func(cmd *cobra.Command, args []string) error {
-			ctx := cmd.Context()
-			d, err := initializeDeploymentSettingsCmd(cmd, stack)
-			if err != nil {
-				return err
-			}
-
-			if d.Deployment == nil {
-				return errors.New("Deployment file not initialized, please run `pulumi deployment settings init` instead")
-			}
-
-			if !yes {
-				confirm, err := askForConfirmation("This action will override the stack's deployment settings, "+
-					"do you want to continue?", d.DisplayOptions.Color, false)
-				if err != nil {
-					return err
-				}
-				if !confirm {
-					return nil
-				}
-			}
-
-			err = d.Backend.UpdateStackDeploymentSettings(ctx, d.Stack, d.Deployment.DeploymentSettings)
-			if err != nil {
-				return err
-			}
-
-			fmt.Println("Done")
-
-			return nil
-		}),
-	}
-
-	cmd.PersistentFlags().StringVarP(
-		&stack, "stack", "s", "",
-		"The name of the stack to operate on. Defaults to the current stack")
-
-	cmd.PersistentFlags().BoolVarP(
-		&yes, "yes", "y", false,
-		"Automatically confirm every confirmation prompt")
-
-	return cmd
-}
-
-func newDeploymentSettingsDestroyCmd() *cobra.Command {
-	var stack string
-	var yes bool
-
-	cmd := &cobra.Command{
-		Use:        "destroy",
-		Aliases:    []string{"down", "dn"},
-		SuggestFor: []string{"delete", "kill", "remove", "rm", "stop"},
-		Args:       cmdutil.ExactArgs(0),
-		Short:      "Delete all the stack's deployment settings",
-		Long:       "",
-		Run: cmdutil.RunFunc(func(cmd *cobra.Command, args []string) error {
-			ctx := cmd.Context()
-			d, err := initializeDeploymentSettingsCmd(cmd, stack)
-			if err != nil {
-				return err
-			}
-
-			if !yes {
-				confirm, err := askForConfirmation("This action will clear the stack's deployment settings, "+
-					"do you want to continue?", d.DisplayOptions.Color, false)
-				if err != nil {
-					return err
-				}
-				if !confirm {
-					return nil
-				}
-			}
-
-			err = d.Backend.DestroyStackDeploymentSettings(ctx, d.Stack)
-			if err != nil {
-				return err
-			}
-
-			fmt.Println("Done")
-
-			return nil
-		}),
-	}
-
-	cmd.PersistentFlags().StringVarP(
-		&stack, "stack", "s", "",
-		"The name of the stack to operate on. Defaults to the current stack")
-
-	cmd.PersistentFlags().BoolVarP(
-		&yes, "yes", "y", false,
-		"Automatically confirm every confirmation prompt")
-
-	return cmd
-}
-
 func newDeploymentSettingsConfigureCmd() *cobra.Command {
 	var stack string
 	var gitSSHPrivateKeyPath string
@@ -542,99 +319,6 @@ func newDeploymentSettingsConfigureCmd() *cobra.Command {
 	cmd.PersistentFlags().StringVarP(
 		&gitSSHPrivateKeyPath, "git-ssh-private-key", "k", "",
 		"Private key path")
-
-	cmd.PersistentFlags().StringVarP(
-		&stack, "stack", "s", "",
-		"The name of the stack to operate on. Defaults to the current stack")
-
-	return cmd
-}
-
-func newDeploymentSettingsEnvCmd() *cobra.Command {
-	var stack string
-
-	var secret bool
-	var remove bool
-
-	cmd := &cobra.Command{
-		Use:   "env <key> [value]",
-		Args:  cmdutil.RangeArgs(1, 2),
-		Short: "Update stack's deployment settings secrets",
-		Long:  "",
-		Run: cmdutil.RunFunc(func(cmd *cobra.Command, args []string) error {
-			ctx := cmd.Context()
-			d, err := initializeDeploymentSettingsCmd(cmd, stack)
-			if err != nil {
-				return err
-			}
-
-			if d.Deployment == nil {
-				return errors.New("Deployment file not initialized, please run `pulumi deployment settings init` instead")
-			}
-
-			var (
-				key   string
-				value string
-			)
-
-			if len(args) > 1 {
-				key = args[0]
-			}
-
-			if len(args) == 2 {
-				if remove {
-					return errors.New("value not supported when removing keys")
-				}
-				value = args[1]
-			}
-
-			if d.Deployment.DeploymentSettings.Operation == nil {
-				d.Deployment.DeploymentSettings.Operation = &apitype.OperationContext{}
-			}
-
-			if d.Deployment.DeploymentSettings.Operation.EnvironmentVariables == nil {
-				d.Deployment.DeploymentSettings.Operation.EnvironmentVariables = make(map[string]apitype.SecretValue)
-			}
-
-			if remove {
-				delete(d.Deployment.DeploymentSettings.Operation.EnvironmentVariables, key)
-			} else {
-				var secretValue *apitype.SecretValue
-				if secret {
-					secretValue, err = d.Backend.EncryptStackDeploymentSettingsSecret(ctx, d.Stack, value)
-					if err != nil {
-						return err
-					}
-				} else {
-					secretValue = &apitype.SecretValue{
-						Value:  value,
-						Secret: false,
-					}
-				}
-
-				d.Deployment.DeploymentSettings.Operation.EnvironmentVariables[key] = *secretValue
-			}
-
-			err = saveProjectStackDeployment(d.Deployment, d.Stack)
-			if err != nil {
-				return err
-			}
-
-			fmt.Println("Done")
-
-			return nil
-		}),
-	}
-
-	cmd.PersistentFlags().BoolVar(
-		&secret, "secret", false,
-		"whether the value should be treated as a secret and be encrypted")
-
-	cmd.PersistentFlags().BoolVar(
-		&remove, "remove", false,
-		"whether the key should be removed")
-
-	cmd.MarkFlagsMutuallyExclusive("secret", "remove")
 
 	cmd.PersistentFlags().StringVarP(
 		&stack, "stack", "s", "",
